Fix build with external libgif and libjson-c
We want to use libtool to build the package to keep the same SONAME.

However, when building with libtool, the --with-gif and --with-libjson-c
configure options require a path prefix to construct the compile and
link command lines correctly. Therefore, the "external" was an invalid
path prefix to link with system-installed libraries.

Finally, dedupe the configure options list.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>